tools/requirements_midjourney.txt
pyautogui>=0.9.54
pygetwindow>=0.0.9
pyperclip>=1.8.2
Pillow>=10.0.0
requests>=2.31.0
